fixed bug in find_EI -- still glitchy at times though
In the while loop, `if len(Es)>0:` needed to be changed to `if Es[-1]
!= 0:` (this is from the original find_EI command line function — it
should reject newly added zero values rather than test for the length
of Es). There are still occasional glitches that usually fix themselves
on the second run, and I’m having a really hard time pinpointing what
these could be. The runtime of the function also seems much longer than
it was before (even with 100 iterations, the bootstrapping takes
forever). So, this is still a work in progress…